Fix 6 bugs + docs updates (pre-VT100 overhaul baseline)
Arp: only mark sequenceDirty when removePadPosition actually found a match,
fold octave transposition with while-loop instead of single subtract,
widen shuffleStepCounter to uint16_t to avoid groove reset at 255.

PotRouter: expand MAX_CC_SLOTS 8→16 (both contexts can be all-CC),
clear stale PB dirty flag on bank switch to prevent wrong-channel send.

ClockManager: only clear timed-out source timestamp (not both — prevents
racing BLE callback), drain external ticks on master mode entry.

MidiEngine: reset AT rate limiter on allNotesOff for instant first-touch,
compare resolved note in flush to catch rapid re-press staleness.

VT100 terminal: use os.read() instead of sys.stdin.buffer.read() for
reliable input on macOS.

Docs: correct USB wire colors, expand USB port documentation.

+++ src/arp/ArpEngine.h
@@ -143,7 +143,7 @@ private:
   //   - pile goes from 0 to 1 note
   //   - pattern change
   // Used as index into groove template: template[counter % 16]
-  uint8_t _shuffleStepCounter;   // Only used modulo 16, wraps at 255 (harmless)
+  uint16_t _shuffleStepCounter;  // Only used modulo 16, uint16_t avoids groove reset at 255
   bool    _tickFlash;            // Set true on each tick(), consumed by LedController
   bool    _waitingForQuantize;  // True after play until quantize boundary reached
   uint8_t _quantizeMode;        // ArpStartMode (0=immediate, 1=beat, 2=bar)
